requirements.txt
bcrypt==3.2.0
boto3==1.18.45
botocore==1.21.45
cffi==1.14.6
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
